Enable `bugprone-implicit-widening-of-multiplication-result` and fix errors
clang-tidy-13 has a new check, `bugprone-implicit-widening-of-multiplication-result`:

"The check diagnoses instances where a result of a multiplication is implicitly widened, and suggests (with fix-it) to either silence the code by making widening explicit, or to perform the multiplication in a wider type, to avoid the widening afterwards."

This seems like a reasonable thing, in theory, so I enabled it and manually fixed everything it complained about (the `-fix` option didn't work well here). Most of the actual changes are probably inconsequential, but checks of this sort seem appropriate for clang-tidy. What do other folks think?

+++ src/CodeGen_D3D12Compute_Dev.cpp
@@ -413,7 +413,7 @@ struct StoragePackUnpack {
     using CodeGen = CodeGen_D3D12Compute_Dev::CodeGen_D3D12Compute_C;
 
     // Shader Model 5.1: threadgroup shared memory is limited 32KB
-    static const size_t ThreadGroupSharedStorageLimit = 32 * 1024;
+    static const size_t ThreadGroupSharedStorageLimit = (size_t)(32 * 1024);
 
     void pack_storage(const Allocate *op, size_t elements, size_t size_in_bytes) {
         // we could try to compact things for smaller types:
